Add database to app and a form to add images Relates #24 #25
