+++ image-a-day-vue/src/App.css
@@ -5,54 +5,3 @@ body {
   font-size: 14px;
   font-family: 'Itim', cursive;
 }
-.gallery {
-  display: flex;
-  justify-content: flex-start;
-  flex-wrap: wrap;
-  width: 95%;
-  margin: 3rem auto;
-}
-.image-box {
-  background-color: #444;
-  margin: 1rem;
-}
-.image {
-  width: 15rem;
-  height: 10rem;
-  padding: 2rem 2rem 0 2rem;
-}
-.text,
-.date {
-  text-align: center;
-  color: #a3a6aa;
-  font-size: 1.2rem;
-}
-.date {
-  font-size: 1rem;
-}
-.image-form {
-  width: 90%;
-  margin: 2rem auto;
-  display: flex;
-  justify-content: space-between;
-}
-.add-title {
-  margin: 2rem;
-}
-.form-input {
-  border: none;
-  background-color: transparent;
-  border-bottom: 1px solid #666;
-  text-align: left;
-  color: #666;
-  line-height: 1.4rem;
-  font-size: 1.2rem;
-  margin: 1rem;
-  padding-left: 1rem;
-}
-.form-input:focus {
-  outline: 0;
-}
-.delete {
-  color: #fff;
-}
