+++ image-a-day-vue/src/App.html
@@ -1,6 +1,6 @@
 
 <div id="app">
-  <!-- <app-header></app-header>
+  <app-header></app-header>
   <router-view/>
-  <app-footer></app-footer> -->
+  <app-footer></app-footer>
 </div>
