+++ image-a-day-vue/src/components/Home/Home.css
@@ -23,3 +23,29 @@
 .date {
   font-size: 1rem;
 }
+.image-form {
+  width: 90%;
+  margin: 2rem auto;
+  display: flex;
+  justify-content: space-between;
+}
+.add-title {
+  margin: 2rem;
+}
+.form-input {
+  border: none;
+  background-color: transparent;
+  border-bottom: 1px solid #666;
+  text-align: left;
+  color: #666;
+  line-height: 1.4rem;
+  font-size: 1.2rem;
+  margin: 1rem;
+  padding-left: 1rem;
+}
+.form-input:focus {
+  outline: 0;
+}
+.delete {
+  color: #fff;
+}
