+++ image-a-day-vue/src/components/Home/Home.vue
@@ -1,85 +1,52 @@
 <template src="./Home.html"></template>
+
 <style scoped lang="css">
   @import 'Home.css'
 </style>
 
 <script>
+import Firebase from 'firebase';
+  let config = {
+    apiKey: "AIzaSyDLQms4Q-ZUdADyWjCVOsWXzfTzsXxjiKo",
+    authDomain: "image-a-day.firebaseapp.com",
+    databaseURL: "https://image-a-day.firebaseio.com",
+    projectId: "image-a-day",
+    storageBucket: "image-a-day.appspot.com",
+    messagingSenderId: "20401690956"
+  }
+
+  let app = Firebase.initializeApp(config);
+  let db = app.database();
+
+  let imagesRef = db.ref('images');
+
   export default {
-    data() {
+    name: 'App',
+    firebase: {
+      images: imagesRef
+    },
+    data () {
       return {
-        images: [
-        { source: "/march/bike.png",
-          text: "Lime Bike in the Back Yard",
-          date: "March 15, 2018",
-          alt: "Bike leaning on a tree in the yard"
-        },
-        { source: "/march/tweet.png",
-          text: "100 Days Of Code",
-          date: "March 16, 2018",
-          alt: "Tweet for joining the 100 days of code coding challenge"
-        },
-        { source: "/march/squirrel.jpg",
-          text: "Squirrel From the Window",
-          date: "March 17, 2018",
-          alt: "A squirrel eats a nut in the front yard"
-        },
-        { source: "/march/lake.png",
-          text: "Potato Creek State Park, IN",
-          date: "March 18, 2018",
-          alt: "A lake at a state park"
-        },
-        { source: "/march/plates.png",
-          text: "Now We Officially Moved to IN",
-          date: "March 19, 2018",
-          alt: "Indiana license plate"
-        },
-        { source: "/march/university.png",
-          text: "On My Way to a Lacture",
-          date: "March 20, 2018",
-          alt: "Notre Dame university building"
-        },
-        { source: "/march/apollo.jpg",
-          text: "The Sun is Finally Out",
-          date: "March 21, 2018",
-          alt: "A dog sitting outside"
-        },
-        { source: "/march/simpson.png",
-          text: "Placeholders Can Be Fun",
-          date: "March 22, 2018",
-          alt: "Homer Simpson as a placeholder image"
-        },
-        { source: "/march/flags.jpg",
-          text: "In front of Notre Dame Library",
-          date: "March 23, 2018",
-          alt: "Flags and the library entrance"
-        },
-        { source: "/march/hackathon.jpg",
-          text: "Hesburgh Library Hackathon",
-          date: "March 24, 2018",
-          alt: "Hackathon T-shirt"
-        },
-        { source: "/march/hackathon-end.jpg",
-          text: "Hesburgh Library Hackathon",
-          date: "March 25, 2018",
-          alt: "People at the hackathon"
-        },
-        { source: "/march/tree.jpg",
-          text: "Just a Tree",
-          date: "March 26, 2018",
-          alt: "tree"
-        },
-        { source: "/march/trello.jpg",
-          text: "My Trello Board",
-          date: "March 27, 2018",
-          alt: "Trello board"
-        },
-        { source: "/march/bitcoin.jpg",
-          text: "First Chrome Extension",
-          date: "March 28, 2018",
-          alt: "Screen shot of a Bitcoin Chrome Extension"
+        newImage: {
+          text: '',
+          source: '',
+          date: '',
+          alt: ''
         }
-        ]
+      }
+    },
+    methods: {
+      addImage: function(){
+        imagesRef.push(this.newImage);
+        this.newImage.text = '';
+        this.newImage.source = '';
+        this.newImage.date = '';
+        this.newImage.alt = '';
+      },
+      removeImage: function(image){
+        imagesRef.child(image['.key']).remove();
       }
     }
   }
+
 </script>
